Added product dao

+++ src/main/java/ru/akirakozov/sd/refactoring/entity/Product.java
@@ -5,15 +5,15 @@ import java.util.Objects;
 public class Product {
     private final Integer id;
     private final String name;
-    private final int price;
+    private final long price;
 
-    public Product(Integer id, String name, int price) {
+    public Product(Integer id, String name, long price) {
         this.id = id;
         this.name = name;
         this.price = price;
     }
 
-    public Product(String name, int price) {
+    public Product(String name, long price) {
         this(null, name, price);
     }
 
@@ -25,7 +25,7 @@ public class Product {
         return name;
     }
 
-    public int getPrice() {
+    public long getPrice() {
         return price;
     }
